refactor(parking-lot): 코드리뷰 내용 반영 #45
