+++ src/main/java/com/parkez/auth/web/TokenController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.parkez.auth.dto.request.RefreshTokenRequest;
 import com.parkez.auth.dto.response.TokenResponse;
 import com.parkez.auth.service.AuthService;
-import com.parkez.common.response.Response;
+import com.parkez.common.dto.response.Response;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
