+++ src/main/java/com/parkez/common/dto/response/DefaultResponse.java
@@ -1,4 +1,4 @@
-package com.parkez.common.response;
+package com.parkez.common.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Getter;
