+++ src/main/java/com/parkez/image/ImageController.java
@@ -1,8 +1,8 @@
 package com.parkez.image;
 
+import com.parkez.common.dto.response.Response;
 import com.parkez.image.dto.request.ImageRequest;
 import com.parkez.image.dto.response.ImageUrlResponse;
-import com.parkez.common.response.Response;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
@@ -34,8 +34,7 @@ public class ImageController {
     public Response<ImageUrlResponse> update(
             @RequestPart("request") ImageRequest request,
             @RequestPart("images") List<MultipartFile> imageFiles
-    ){
-        ;
+    ) {
         return Response.of(imageService.replace(request, imageFiles));
     }
